// store/actions/types.ts
export enum ActionType {
  NAVIGATION = 'NAVIGATION', 
  API = 'API' 
}

export enum AuthenticationActionType {
  AUTHORISE_USER = 'AUTHORISE_USER',
  AUTHORISE_ERROR = 'AUTHORISE_ERROR',
  SIGN_UP_SUCCESS = 'SIGN_UP_SUCCESS',
  SIGN_UP_FAIL = 'SIGN_UP_FAIL',
  SIGN_IN_SUCCESS = 'SIGN_IN_SUCCESS',
  SIGN_IN_FAIL = 'SIGN_IN_FAIL',
  SIGN_OUT_SUCCESS = 'SIGN_OUT_SUCCESS',
  SIGN_OUT_FAIL = 'SIGN_OUT_FAIL'
}